Split core/ui for ReturnTypeSubProcessor
