+++ org.eclipse.jdt.core.manipulation/proposals/org/eclipse/jdt/internal/ui/text/correction/proposals/LinkedCorrectionProposalCore.java
@@ -24,7 +24,7 @@ import org.eclipse.jdt.internal.corext.fix.LinkedProposalPositionGroupCore.Propo
 
 import org.eclipse.jdt.ui.text.java.correction.ASTRewriteCorrectionProposalCore;
 
-public class LinkedCorrectionProposalCore extends ASTRewriteCorrectionProposalCore {
+public class LinkedCorrectionProposalCore extends ASTRewriteCorrectionProposalCore implements ILinkedCorrectionProposalCore {
 	public LinkedProposalModelCore fLinkedProposalModel;
 
 	public LinkedCorrectionProposalCore(String name, ICompilationUnit cu, ASTRewrite rewrite, int relevance) {
